# docs/swagger.yaml
definitions:
...
  casts.Token:
    properties:
      expired_at:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token:
        type: string
      token_type:
        type: string
    type: object
...
  helpers.ResponseParams-models_Product:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      errors:
        additionalProperties:
          type: string
        type: object
      item:
        $ref: '#/definitions/models.Product'
      message:
        type: string
      reference:
        type: string
      status:
        type: string
      token:
        $ref: '#/definitions/casts.Token'
      total:
        type: integer
    type: object
...
  models.Category:
    properties:
      category:
        type: string
      created_at:
        type: string
      id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Product:
    properties:
      category:
        allOf:
        - $ref: '#/definitions/models.Category'
        description: Store    *Store    `json:"store"`
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      margin:
        type: number
      name:
        type: string
      price:
        type: number
      received_at:
        type: string
      reference:
        type: string
      sold:
        type: integer
      stock:
        type: integer
      store_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.User:
    properties:
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      pin:
        type: string
      reference:
        type: string
      username:
        type: string
    type: object
  requests.CategoryRequest:
    properties:
      category:
        example: Electronics
        type: string
      id:
        type: integer
    required:
    - category
    type: object
...
  requests.ProductRequest:
    properties:
      category_id:
        example: 2
        type: integer
      description:
        example: A brief description of the product
        type: string
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      margin:
        example: 10
        type: number
      name:
        example: Product Name
        type: string
      price:
        example: 99.99
        type: number
      received_at:
        example: "2023-10-10T00:00:00Z"
        type: string
      reference:
        example: PRD001
        type: string
      sold:
        example: 10
        type: integer
      stock:
        example: 100
        type: integer
      store_id:
        example: 1
        type: integer
    required:
    - category_id
    - name
    - price
    - store_id
    type: object
...
info:
  contact: {}
paths:
...
  /categories:
    get:
      description: Retrieve a list of all categories, including related products
      produces:
      - application/json
      responses:
        "200":
          description: List of categories with products
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List all categories
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Create a new category or update an existing one by ID
      parameters:
      - description: Category Data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/requests.CategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Created or updated category
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid input data
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create or update a category
      tags:
      - categories
  /categories/{id}:
    delete:
      description: Delete a specific category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Category not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete a category by ID
      tags:
      - categories
    get:
      description: Retrieve a category by its ID, including related products
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category with products
          schema:
            $ref: '#/definitions/models.Category'
        "404":
          description: Category not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get a category by ID
      tags:
      - categories
  /file/{key}/{filename}:
    get:
      consumes:
      - application/json
      description: Serve file
      parameters:
      - description: Signature
        in: query
        name: signature
        type: string
      - description: File key
        in: path
        name: key
        required: true
        type: string
      - description: File path
        in: path
        name: path
        required: true
        type: string
      produces:
      - image/jpeg
      responses:
        "200":
          description: File
          schema:
            type: string
      summary: Serve file
      tags:
      - File
  /file/public/{key}/{filename}:
    get:
      consumes:
      - application/json
      description: Serve file directly without JWT authentication
      parameters:
      - description: File key
        in: path
        name: key
        required: true
        type: string
      - description: File name
        in: path
        name: filename
        required: true
        type: string
      produces:
      - image/jpeg
      responses:
        "200":
          description: File
          schema:
            type: file
        "404":
          description: File not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Serve file without authentication
      tags:
      - File
...
  /products:
    get:
      consumes:
      - application/json
      description: API untuk mendapatkan semua produk
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: offset
        type: integer
      - in: query
        name: order_by
        type: string
      - enum:
        - asc
        - desc
        in: query
        name: order_direction
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ResponseParams-models_Product'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Product'
                  type: array
              type: object
      summary: Get all products
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: API untuk membuat atau mengupdate produk
      parameters:
      - description: Product request body
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/requests.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ResponseParams-models_Product'
            - properties:
                item:
                  $ref: '#/definitions/models.Product'
              type: object
      summary: Create/Update product
      tags:
      - Product
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: API untuk menghapus produk
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseParams-models_Product'
      summary: Delete product
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: API untuk mendapatkan produk berdasarkan ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ResponseParams-models_Product'
            - properties:
                item:
                  $ref: '#/definitions/models.Product'
              type: object
      summary: Get product by ID
      tags:
      - Product
...
  /users:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: Show all users
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User object
        in: body
        name: JSON
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Upsert a user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: User deleted
          schema:
            type: string
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Show a user
      tags:
      - users
security:
- X-Api-Key: []
- Bearer: []
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
  X-Api-Key:
    in: header
    name: X-Api-Key
    type: apiKey
swagger: "2.0"